+++ src/components/button/button_icon/button_icon.tsx
@@ -46,7 +46,6 @@ export type EuiButtonIconColor =
   | 'subdued'
   | 'success'
   | 'text'
-  | 'secondary'
   | 'warning';
 
 export interface EuiButtonIconProps extends CommonProps {
@@ -96,7 +95,6 @@ const colorToClassNameMap: { [color in EuiButtonIconColor]: string } = {
   success: 'euiButtonIcon--success',
   text: 'euiButtonIcon--text',
   warning: 'euiButtonIcon--warning',
-  secondary: 'euiButtonIcon--secondary',
 };
 
 export const COLORS = keysOf(colorToClassNameMap);