+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
 - Upgraded `highlight.js` to 9.18.5 ([#4313](https://github.com/elastic/eui/pull/4313)) 
-- Added `secondary` and `warning` colors to `EuiButtonEmpty` and `secondary` to `EuiButtonIcon` ([#4325](https://github.com/elastic/eui/pull/4325))
+- Added `success` and `warning` colors to `EuiButtonEmpty` ([#4325](https://github.com/elastic/eui/pull/4325))
 
 
 **Theme: Amsterdam**